+++ NEWS.md
@@ -1,2 +1,9 @@
+- 0.2.2:
+    - 'init' is done better using `data.table`. Time reduced to less than 1/5 what is used to take.
+    - 'init' has the new default of `local = FALSE`.
+
+- 0.2.1:
+    - fixed the 'plot', so that it does not mask the plot generic
+
 - 0.2.0:
     - known issue with legend not appearning in knitr output